// packages/devextreme/js/__internal/viz/core/types.ts
export interface Coords {
  x: number;
  y: number;
}

export interface Size {
  width: number;
  height: number;
}

export interface BBox {
  x: number;
  y: number;
  width: number;
  height: number;
}

export interface Bounds {
  left: number;
  top: number;
  right: number;
  bottom: number;
}

export interface Canvas {
  left: number;
  top: number;
  right: number;
  bottom: number;
  width: number;
  height: number;
}